Fix vendor time Fix armor resist not affecting damage Reduce lair max dam Fix harvesting typo to show correct extracted Add missing xp modifiers Fix credit split on mission Fix repair to properly repair

+++ MMOCoreORB/src/server/zone/objects/tangible/wearables/WearableObjectImplementation.cpp
@@ -389,7 +389,7 @@ String WearableObjectImplementation::repairAttempt(int repairChance) {
 		repairChance = 0;
 	}
 
-	int reduction = (float)(Math::min(95, Math::max(75, repairChance)) / 100.f);
+	float reduction = (float)(Math::min(95, Math::max(75, repairChance)) / 100.f);
 
 	setMaxCondition(getMaxCondition() * reduction, true);
 	setConditionDamage(0, true);
